Load multiple levels (#9)
* feat: Load multi-level txt file and select a single level from it

No cli for this yet so you have to set the filename and level number (0
indexed) in main.cpp.

* feat: Allow file and level selection through CLI

Basic CLI to select the level file and optionally the level. This is
quite fragile e.g. it will allow you to select levels that don't exist
without complaining, but for testing it will do for now.

* build: Update Makefile run target to use new CLI

* chore: clean up level files

test level file "level.txt" no longer needed as we can load actual level
packs now. Not sure about redistributing the microban txt file so to be
safe I'm excluding it from the repo.